ua_bot-main/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.7.0
boto3==1.34.62
botocore==1.34.62
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.5.14
distlib==0.3.8
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.0.2
flask-crontab==0.1.2
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
gradio==3.44.2
gradio_client==0.5.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
identify==2.5.35
idna==3.6
importlib_resources==6.3.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.0.287
langsmith==0.0.92
Levenshtein==0.25.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
multidict==6.0.5
mypy-extensions==1.0.0
nodeenv==1.8.0
numexpr==2.9.0
numpy==1.26.4
openai==0.28.0
orjson==3.9.15
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
s3transfer==0.10.0
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.5.1
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.7
uvicorn==0.28.0
virtualenv==20.25.1
websockets==11.0.3
Werkzeug==3.0.1
yarl==1.9.4
